# Speed up diff stat in manpages