The diff stat is now done by raw analysis of diff command output. For
instance for git-config man page we switch from:

    Completed 200 OK in 7349ms (Views: 7148.4ms | ActiveRecord: 134.3ms |
 Allocations: 5879568)

to:

    Completed 200 OK in 737ms (Views: 590.1ms | ActiveRecord: 139.2ms |
 Allocations: 658373)

which is a 10x speed up