@jnavila 2021-01-09
Well, I played a bit with how to render the diffstat. The performance is still as good.