@peff 2021-01-11
This looks good to me on the pr-deployment, so I'm going to merge, and we'll keep an eye on it. I think the performance improvement should be enough that we don't need to worry anymore, but if it bites us again, we can think about something like #1564. Thanks @jnavila for coming up with this solution.