@peff 2021-01-11
> I think the performance improvement should be enough that we don't need to worry anymore,

I flushed the redis/rails caches and did a direct hit on `git-scm.herokuapp.com/docs/git-config`, and it seemed to come up pretty promptly. So that's a good sign.

One curiosity is that the page seems to show mostly deletions between versions, rather than additions. Which seems intuitively wrong, as we tend to add new config options rather than deleting them. I wonder if we have the order of the diff messed up.