@peff 2021-01-11
> One curiosity is that the page seems to show mostly deletions between versions, rather than additions. Which seems intuitively wrong, as we tend to add new config options rather than deleting them. I wonder if we have the order of the diff messed up.

Your code behaves the same as the original. If anything, I would suggest the problem is in the caller https://github.com/git/git-scm.com/blob/d276fcdeae3ec8e88211d292e091356fa4366ca9/app/models/doc_file.rb#L63

which is instructing us to diff from (for example) v2.30.0 to v2.29.0. I suppose one could make an argument that we are looking backwards in time, and showing the changes as we move back. But it is definitely not what I expected (the fact that nobody has mentioned this before also fuels my notion that these diffstats are not actually useful :wink:).