@jnavila 2021-01-11
> One curiosity is that the page seems to show mostly deletions between versions, rather than additions. Which seems intuitively wrong, as we tend to add new config options rather than deleting them. I wonder if we have the order of the diff messed up.

I'm pretty sure it worked correctly in my dev env. I'll check again tonight.